requirements.txt
haystack-ai==2.5.0
Flask==3.0.3
requests==2.32.3